+++ packages/http-client-csharp/generator/Microsoft.TypeSpec.Generator.ClientModel/src/Providers/RestClientProvider.cs
@@ -115,8 +115,10 @@ namespace Microsoft.TypeSpec.Generator.ClientModel.Providers
                     Declare("request", message.Request().ToApi<HttpRequestApi>(), out HttpRequestApi request),
                     request.SetMethod(operation.HttpMethod),
                     Declare("uri", New.Instance(request.UriBuilderType), out ScopedApi uri),
-                    uri.Reset(ClientProvider.EndpointField).Terminate(),
-                    .. AppendPathParameters(uri, operation, paramMap),
+                    operation.Paging?.NextLink != null ?
+                        uri.Reset(ScmKnownParameters.NextPage.AsExpression().NullCoalesce(ClientProvider.EndpointField)).Terminate() :
+                        uri.Reset(ClientProvider.EndpointField).Terminate(),
+                    .. ConditionallyAppendPathParameters(operation, uri, paramMap),
                     .. AppendQueryParameters(uri, operation, paramMap),
                     request.SetUri(uri),
                     .. AppendHeaderParameters(request, operation, paramMap),
@@ -125,6 +127,22 @@ namespace Microsoft.TypeSpec.Generator.ClientModel.Providers
                     Return(message)
                 ]),
                 this);
+        }
+
+        private IReadOnlyList<MethodBodyStatement> ConditionallyAppendPathParameters(InputOperation operation, ScopedApi uri, Dictionary<string, ParameterProvider> paramMap)
+        {
+            if (operation.Paging?.NextLink != null)
+            {
+                return
+                [
+                    new IfStatement(ScmKnownParameters.NextPage.Equal(Null))
+                    {
+                        new MethodBodyStatements([..AppendPathParameters(uri, operation, paramMap)])
+                    }
+                ];
+            }
+
+            return AppendPathParameters(uri, operation, paramMap);
         }
 
         private IReadOnlyList<MethodBodyStatement> GetSetContent(HttpRequestApi request, IReadOnlyList<ParameterProvider> parameters)
@@ -186,7 +204,7 @@ namespace Microsoft.TypeSpec.Generator.ClientModel.Providers
 
             foreach (var inputParameter in operation.Parameters)
             {
-                if (inputParameter.Location != RequestLocation.Header)
+                if (inputParameter.Location != InputRequestLocation.Header)
                     continue;
 
                 CSharpType? type;
@@ -216,7 +234,7 @@ namespace Microsoft.TypeSpec.Generator.ClientModel.Providers
 
             foreach (var inputParameter in operation.Parameters)
             {
-                if (inputParameter.Location != RequestLocation.Query)
+                if (inputParameter.Location != InputRequestLocation.Query)
                     continue;
 
                 string? format;
@@ -321,7 +339,7 @@ namespace Microsoft.TypeSpec.Generator.ClientModel.Providers
             return new IfStatement(valueExpression.NotEqual(Null)) { originalStatement };
         }
 
-        private IEnumerable<MethodBodyStatement> AppendPathParameters(ScopedApi uri, InputOperation operation, Dictionary<string, ParameterProvider> paramMap)
+        private IReadOnlyList<MethodBodyStatement> AppendPathParameters(ScopedApi uri, InputOperation operation, Dictionary<string, ParameterProvider> paramMap)
         {
             Dictionary<string, InputParameter> inputParamHash = new(operation.Parameters.ToDictionary(p => p.Name));
             List<MethodBodyStatement> statements = new(operation.Parameters.Count);
@@ -370,7 +388,7 @@ namespace Microsoft.TypeSpec.Generator.ClientModel.Providers
                 else
                 {
                     inputParam = inputParamHash[paramName];
-                    if (inputParam.Location == RequestLocation.Path || inputParam.Location == RequestLocation.Uri)
+                    if (inputParam.Location == InputRequestLocation.Path || inputParam.Location == InputRequestLocation.Uri)
                     {
                         GetParamInfo(paramMap, operation, inputParam, out type, out format, out valueExpression);
                     }
@@ -447,7 +465,7 @@ namespace Microsoft.TypeSpec.Generator.ClientModel.Providers
                     property.Summary,
                     property.Doc,
                     property.Type,
-                    RequestLocation.Body,
+                    InputRequestLocation.Body,
                     null,
                     InputOperationParameterKind.Method,
                     property.IsRequired,
@@ -472,7 +490,7 @@ namespace Microsoft.TypeSpec.Generator.ClientModel.Providers
 
         private static bool TryGetSpecialHeaderParam(InputParameter inputParameter, [NotNullWhen(true)] out ParameterProvider? parameterProvider)
         {
-            if (inputParameter.Location == RequestLocation.Header)
+            if (inputParameter.Location == InputRequestLocation.Header)
             {
                 return _knownSpecialHeaderParams.TryGetValue(inputParameter.NameInRequest, out parameterProvider);
             }
@@ -510,7 +528,8 @@ namespace Microsoft.TypeSpec.Generator.ClientModel.Providers
         internal static List<ParameterProvider> GetMethodParameters(InputOperation operation, MethodType methodType)
         {
             SortedList<int, ParameterProvider> sortedParams = [];
-            int path = 0;
+            int paging = 0;
+            int path = 1;
             int required = 100;
             int bodyRequired = 200;
             int bodyOptional = 300;
@@ -519,9 +538,12 @@ namespace Microsoft.TypeSpec.Generator.ClientModel.Providers
 
             foreach (InputParameter inputParam in operation.Parameters)
             {
-                if ((inputParam.Kind != InputOperationParameterKind.Method && inputParam.Kind != InputOperationParameterKind.Spread)
-                    || TryGetSpecialHeaderParam(inputParam, out var _))
+                if ((inputParam.Kind != InputOperationParameterKind.Method &&
+                     inputParam.Kind != InputOperationParameterKind.Spread) ||
+                    TryGetSpecialHeaderParam(inputParam, out _))
+                {
                     continue;
+                }
 
                 var spreadInputModel = inputParam.Kind == InputOperationParameterKind.Spread ? GetSpreadParameterModel(inputParam) : null;
 
@@ -529,7 +551,7 @@ namespace Microsoft.TypeSpec.Generator.ClientModel.Providers
 
                 if (methodType is MethodType.Protocol or MethodType.CreateRequest)
                 {
-                    if (inputParam.Location == RequestLocation.Body)
+                    if (inputParam.Location == InputRequestLocation.Body)
                     {
                         if (methodType == MethodType.CreateRequest)
                         {
@@ -601,12 +623,18 @@ namespace Microsoft.TypeSpec.Generator.ClientModel.Providers
                 sortedParams.Add(bodyRequired++, ScmKnownParameters.ContentType);
             }
 
-            // All the parameters should be required for the CreateRequest method
             if (methodType == MethodType.CreateRequest)
             {
+                // All the parameters should be required for the CreateRequest method
                 foreach (var parameter in sortedParams.Values)
                 {
                     parameter.DefaultValue = null;
+                }
+
+                if (operation.Paging?.NextLink != null)
+                {
+                    // Next link operations will always have an endpoint parameter in the CreateRequest method
+                    sortedParams.Add(paging, ScmKnownParameters.NextPage);
                 }
             }