+++ packages/http-client-java/generator/http-client-generator-test/src/main/java/tsptest/builtin/BuiltinClient.java
@@ -92,6 +92,9 @@ public final class BuiltinClient {
      *         unixTimestamp: Long (Optional)
      *         base64: byte[] (Optional)
      *         base64url: Base64Url (Optional)
+     *         unknownDurationFormat: String (Optional)
+     *         unknownDateTimeFormat: String (Optional)
+     *         unknownBytes: String (Optional)
      *     }
      * }
      * }
@@ -150,6 +153,9 @@ public final class BuiltinClient {
      *         unixTimestamp: Long (Optional)
      *         base64: byte[] (Optional)
      *         base64url: Base64Url (Optional)
+     *         unknownDurationFormat: String (Optional)
+     *         unknownDateTimeFormat: String (Optional)
+     *         unknownBytes: String (Optional)
      *     }
      * }
      * }