+++ packages/http-server-csharp/src/lib/utils.ts
@@ -112,8 +112,6 @@ export function getCSharpType(
       else return { type: standardScalars.get("numeric")!, value: new NumericValue(enumValue) };
     case "Intrinsic":
       return getCSharpTypeForIntrinsic(program, type);
-    case "Object":
-      return { type: UnknownType };
     case "ModelProperty":
       return getCSharpType(program, type.type, namespace);
     case "Scalar":
@@ -173,6 +171,16 @@ export function getCSharpType(
           }),
         };
       }
+      if (isRecord(type))
+        return {
+          type: new CSharpType({
+            name: "JsonObject",
+            namespace: "System.Text.Json.Nodes",
+            isBuiltIn: false,
+            isValueType: false,
+            isClass: false,
+          }),
+        };
       let name: string = type.name;
       if (isTemplateInstance(type)) {
         name = getModelInstantiationName(program, type, name);
@@ -443,20 +451,25 @@ export function formatComment(
 ): string {
   function getNextLine(target: string): string {
     for (let i = lineLength - 1; i > 0; i--) {
-      if ([" ", ".", "?", ",", ";"].includes(target.charAt(i))) {
-        return `/// ${text.substring(0, i).replaceAll("\n", " ")}`;
+      if ([" ", ";"].includes(target.charAt(i))) {
+        return `${target.substring(0, i)}`;
+      }
+    }
+    for (let i = lineLength - 1; i < target.length; i++) {
+      if ([" ", ";"].includes(target.charAt(i))) {
+        return `${target.substring(0, i)}`;
       }
     }
 
-    return `/// ${text.substring(0, lineLength)}`;
+    return `${target.substring(0, lineLength)}`;
   }
-  let remaining: string = text;
+  let remaining: string = text.replaceAll("\n", " ");
   const lines: string[] = [];
   while (remaining.length > lineLength) {
     const currentLine = getNextLine(remaining);
     remaining =
       remaining.length > currentLine.length ? remaining.substring(currentLine.length + 1) : "";
-    lines.push(currentLine);
+    lines.push(`/// ${currentLine}`);
   }
 
   if (remaining.length > 0) lines.push(`/// ${remaining}`);
@@ -550,11 +563,7 @@ export function ensureCSharpIdentifier(
       location = `union ${target.name}`;
       break;
     case "UnionVariant": {
-      if (target.node !== undefined) {
-        const parent = program.checker.getTypeForNode(target.node.parent!);
-        if (parent?.kind === "Union")
-          location = `variant ${String(target.name)} in union ${parent?.name}`;
-      }
+      location = `variant ${String(target.name)} in union ${target.union.name}`;
       break;
     }
   }
@@ -708,6 +717,7 @@ export class HttpMetadata {
     switch (responseType.kind) {
       case "Model":
         if (responseType.indexer && responseType.indexer.key.name !== "string") return responseType;
+        if (isRecord(responseType)) return responseType;
         const bodyProp = new ModelInfo().filterAllProperties(
           program,
           responseType,
@@ -1274,8 +1284,6 @@ export class CSharpOperationHelpers {
           defaultValue: `[${defaults.join(", ")}]`,
           nullableType: csharpType.isNullable,
         };
-      case "Object":
-        return { typeReference: code`object`, defaultValue: undefined, nullableType: false };
       case "Model":
         let modelResult: EmittedTypeInfo;
         const cachedResult = this.#anonymousModels.get(tsType);
@@ -1283,7 +1291,10 @@ export class CSharpOperationHelpers {
           return cachedResult;
         }
         if (isRecord(tsType)) {
-          modelResult = { typeReference: code`JsonObject`, nullableType: false };
+          modelResult = {
+            typeReference: code`System.Text.Json.Nodes.JsonObject`,
+            nullableType: false,
+          };
         } else {
           modelResult = {
             typeReference: code`${this.emitter.emitTypeReference(tsType)}`,