+++ packages/html-program-viewer/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log - @typespec/html-program-viewer
 
+## 0.66.0
+
+### Features
+
+- [#5996](https://github.com/microsoft/typespec/pull/5996) Emitter Framework V2
+
+
 ## 0.65.0
 
 ### Bump dependencies